pyproject.toml
[tool.poetry]
name = "nonebot_plugin_alias"
version = "0.1.0"
description = "A simple plugin for adding aliases for Nonebot command"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-alpha.13"
nonebot-adapter-cqhttp = "^2.0.0-alpha.13"

[tool.poetry.dev-dependencies]
nb-cli = "^0.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
